+++ src/pandemia/world/world_factory/abmlux_world_factory.py
@@ -53,7 +53,7 @@ class ABMluxWorldFactory(WorldFactory):
         assert number_of_regions == 1
 
         world.travel_matrix =\
-            np.zeros((number_of_regions, number_of_regions), dtype=np.int64)
+            np.zeros((number_of_regions, number_of_regions), dtype=np.int32)
 
         return world
