+++ src/pandemia/world/world_factory/control_world_factory.py
@@ -157,10 +157,10 @@ class ControlWorldFactory(WorldFactory):
 
         # This will be the matrix returned by the function
         baseline_agents_travelling_matrix =\
-            np.zeros((num_of_regions, num_of_regions), dtype=np.int64)
+            np.zeros((num_of_regions, num_of_regions), dtype=np.int32)
 
         # Calculate local travel, rescaled according to step size
-        share_matrix = np.zeros((num_of_regions, num_of_regions), dtype=np.float64)
+        share_matrix = np.zeros((num_of_regions, num_of_regions), dtype=float)
         for region in regions:
             for other_region in regions:
                 if region.id != other_region.id:
